Move API modules into sub-folders
